+++ src/main/java/ru/mail/polis/ads/part1/blinkyz/FastScanner.java
@@ -1,0 +1,31 @@
+package ru.mail.polis.ads.part1.blinkyz;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.util.StringTokenizer;
+
+class FastScanner {
+    private final BufferedReader reader;
+    private StringTokenizer tokenizer;
+
+    FastScanner(final InputStream in) {
+        reader = new BufferedReader(new InputStreamReader(in));
+    }
+
+    String next() {
+        while (tokenizer == null || !tokenizer.hasMoreTokens()) {
+            try {
+                tokenizer = new StringTokenizer(reader.readLine());
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+        return tokenizer.nextToken();
+    }
+
+    int nextInt() {
+        return Integer.parseInt(next());
+    }
+}
