+++ src/main/java/ru/mail/polis/ads/part1/blinkyz/Problem1.java
@@ -1,7 +1,6 @@
 package ru.mail.polis.ads.part1.blinkyz;
 
-import java.io.*;
-import java.util.StringTokenizer;
+import java.io.PrintWriter;
 
 public class Problem1 {
     private Problem1() {
@@ -9,32 +8,8 @@ public class Problem1 {
     }
 
     private static void solve(final FastScanner in, final PrintWriter out) {
-        int num = in.nextInt();
+        final int num = in.nextInt();
         System.out.println(num / 10 + " " + num % 10);
-    }
-
-    private static class FastScanner {
-        private final BufferedReader reader;
-        private StringTokenizer tokenizer;
-
-        FastScanner(final InputStream in) {
-            reader = new BufferedReader(new InputStreamReader(in));
-        }
-
-        String next() {
-            while (tokenizer == null || !tokenizer.hasMoreTokens()) {
-                try {
-                    tokenizer = new StringTokenizer(reader.readLine());
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-            return tokenizer.nextToken();
-        }
-
-        int nextInt() {
-            return Integer.parseInt(next());
-        }
     }
 
     public static void main(final String[] arg) {
