+++ src/main/java/ru/mail/polis/ads/part1/blinkyz/Problem1618.java
@@ -1,14 +1,12 @@
 package ru.mail.polis.ads.part1.blinkyz;
 
-import java.io.*;
+import java.io.PrintWriter;
 import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.StringTokenizer;
 
 /**
  * Problem 1618.
- * <p>
  * Link: {@code https://www.e-olymp.com/ru/problems/1618}.
  * Tests: {@code https://www.e-olymp.com/ru/submissions/5736502}.
  */
@@ -20,10 +18,12 @@ public class Problem1618 {
     /**
      * Subsequence struct.
      *
-     * @implNote We store only size and last char entrance position for the subsequence as we do not need to access the whole subsequence
-     * but only last position as we just want to know if we can insert a new element into this subsequence.
-     * Also we need to know size of the subsequence to know if there's benefit from changing last entrance pos to the new one, that is we
-     * update subsequence's last pos only and only if {@code (new entrance pos)} < {@code lastPos}.
+     * @implNote We store only size and last char entrance position for the subsequence as we do not need to access
+     * the whole subsequence but only last position as we just want to know if we can insert a new element
+     * into this subsequence.
+     * Also we need to know size of the subsequence to know if there's benefit from changing last
+     * entrance pos to the new one, that is we update subsequence's last pos only and
+     * only if {@code (new entrance pos)} < {@code lastPos}.
      */
     private static final class Subsequence {
         /**
@@ -62,10 +62,7 @@ public class Problem1618 {
      * @return new subsequences in which we've inserted one of the entrances
      */
     private static List<Subsequence> insert(final LinkedList<Subsequence> subsequences, final List<Integer> entrances) {
-        List<Subsequence> subsToAdd = new ArrayList<>();
-        if (entrances.isEmpty()) {
-            return subsToAdd;
-        }
+        final List<Subsequence> subsToAdd = new ArrayList<>();
 
         // we DO NOT want to create a new subsequence with a greater entrance pos than previous, if current size is
         // the same as with previous entrance
@@ -75,31 +72,23 @@ public class Problem1618 {
         int lastInsertedSize = -1;
         // why would we try to insert current char into the subsequence if we did it already with a smaller entrance
         // pos of the same char?
-        boolean[] isInsertedInCurrentSub = new boolean[subsequences.size()];
-        for (int curEntrancePos : entrances) {
+        final boolean[] isInsertedInCurrentSub = new boolean[subsequences.size()];
+        for (final int curEntrancePos : entrances) {
             for (int j = subsequences.size() - 1; j >= 0; --j) {
                 // is there already was insertion into this subsequence?
                 // if true, we don't want to try insert current entrance as it is greater than previous and there's
                 // no benefit from
                 // doing it
                 if (!isInsertedInCurrentSub[j]) {
-                    Subsequence curSub = subsequences.get(j);
+                    final Subsequence curSub = subsequences.get(j);
                     // as said we can insert a new element only and only if new entrance pos < lastPos
-                    if (curSub.lastPos < curEntrancePos) {
-                        int newSize = curSub.size + 1;
-                        // okay, suppose we've inserted this element into this subsequence and got a new subsequence with
-                        // a bigger size
-                        // but have've got a subsequence of such size before? if true, it means that we're trying to harm
-                        // ourselves as
-                        // we get a subsequence of the same size but with the GREATER last pos
-                        // as said, we want to create a new subsequence only and only if size is bigger
-                        if (newSize > lastInsertedSize) {
-                            Subsequence newSub = new Subsequence(newSize, curEntrancePos);
-                            subsToAdd.add(newSub);
-                            lastInsertedSize = newSize;
-                            isInsertedInCurrentSub[j] = true;
-                            break;
-                        }
+                    if (curSub.lastPos < curEntrancePos && (curSub.size + 1) > lastInsertedSize) {
+                        final int newSize = curSub.size + 1;
+                        final Subsequence newSub = new Subsequence(newSize, curEntrancePos);
+                        subsToAdd.add(newSub);
+                        lastInsertedSize = newSize;
+                        isInsertedInCurrentSub[j] = true;
+                        break;
                     }
                 }
             }
@@ -123,7 +112,7 @@ public class Problem1618 {
         }
 
         for (int i = 0; i < nSize; i++) {
-            int curChar = arr1.get(i); // current char
+            final int curChar = arr1.get(i); // current char
             final List<Integer> entrances = new ArrayList<>(); // all entrances of the current char in the other sequence
             for (int ii = 0; ii < arr2.size(); ii++) {
                 if (arr2.get(ii) == curChar) {
@@ -160,7 +149,7 @@ public class Problem1618 {
 
     private static void insertNewSubsequence(LinkedList<Subsequence> subsequences, Subsequence newSub) {
         for (int i = 0; i < subsequences.size(); i++) {
-            Subsequence curSub = subsequences.get(i);
+            final Subsequence curSub = subsequences.get(i);
             if (curSub.size == newSub.size) {
                 if (newSub.lastPos < curSub.lastPos) {
                     subsequences.set(i, newSub);
@@ -170,30 +159,6 @@ public class Problem1618 {
         }
         // we have a subsequence with a greater size than all of the present or we have empty subsequences list
         subsequences.add(newSub);
-    }
-
-    private static class FastScanner {
-        private final BufferedReader reader;
-        private StringTokenizer tokenizer;
-
-        FastScanner(final InputStream in) {
-            reader = new BufferedReader(new InputStreamReader(in));
-        }
-
-        String next() {
-            while (tokenizer == null || !tokenizer.hasMoreTokens()) {
-                try {
-                    tokenizer = new StringTokenizer(reader.readLine());
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-            return tokenizer.nextToken();
-        }
-
-        int nextInt() {
-            return Integer.parseInt(next());
-        }
     }
 
     public static void main(final String[] arg) {
